requirements.txt
Appium-Python-Client==5.1.1
pytest==8.3.5
allure-pytest==2.14.2
selene==2.0.0rc9
allure-python-commons~=2.14.2
python-dotenv~=1.1.0
requests
pydantic~=2.11.5
selenium~=4.33.0